requirements.txt
boto3>=1.26
pandas>=1.3
numpy>=1.19
scikit-learn>=1.0
matplotlib>=3.3
joblib>=1.1
snowflake-connector-python>=3.0
snowflake-sqlalchemy>=1.4
apache-airflow==2.6.3  # use compatible version for your environment; change if needed
flask>=2.0
python-dotenv>=0.21
